setup base reporter to reduce code repetition
