+++ reporter/otlp_reporter.go
@@ -12,7 +12,6 @@ import (
 
 	lru "github.com/elastic/go-freelru"
 	log "github.com/sirupsen/logrus"
-	"github.com/zeebo/xxh3"
 	"go.opentelemetry.io/collector/pdata/pprofile"
 	"go.opentelemetry.io/collector/pdata/pprofile/pprofileotlp"
 	semconv "go.opentelemetry.io/otel/semconv/v1.25.0"
@@ -31,6 +30,8 @@ var _ Reporter = (*OTLPReporter)(nil)
 
 // OTLPReporter receives and transforms information to be OTLP/profiles compliant.
 type OTLPReporter struct {
+	*BaseReporter
+
 	config *Config
 	// name is the ScopeProfile's name.
 	name string
@@ -53,15 +54,6 @@ type OTLPReporter struct {
 
 	// hostmetadata stores metadata that is sent out with every request.
 	hostmetadata *lru.SyncedLRU[string, string]
-
-	// cgroupv2ID caches PID to container ID information for cgroupv2 containers.
-	cgroupv2ID *lru.SyncedLRU[libpf.PID, string]
-
-	// pdata holds the generator for the data being exported.
-	pdata *pdata.Pdata
-
-	// traceEvents stores reported trace events (trace metadata with frames and counts)
-	traceEvents xsync.RWMutex[map[samples.TraceAndMetaKey]*samples.TraceEvents]
 
 	// pkgGRPCOperationTimeout sets the time limit for GRPC requests.
 	pkgGRPCOperationTimeout time.Duration
@@ -108,7 +100,14 @@ func NewOTLP(cfg *Config) (*OTLPReporter, error) {
 	}
 
 	return &OTLPReporter{
-		config:        cfg,
+		BaseReporter: &BaseReporter{
+			cfg:        cfg,
+			pdata:      data,
+			cgroupv2ID: cgroupv2ID,
+			traceEvents: xsync.NewRWMutex(
+				map[samples.TraceAndMetaKey]*samples.TraceEvents{},
+			),
+		},
 		name:          cfg.Name,
 		version:       cfg.Version,
 		kernelVersion: cfg.KernelVersion,
@@ -121,66 +120,11 @@ func NewOTLP(cfg *Config) (*OTLPReporter, error) {
 		pkgGRPCOperationTimeout: cfg.GRPCOperationTimeout,
 		client:                  nil,
 		rpcStats:                NewStatsHandler(),
-		pdata:                   data,
 		hostmetadata:            hostmetadata,
-		traceEvents: xsync.NewRWMutex(
-			map[samples.TraceAndMetaKey]*samples.TraceEvents{},
-		),
-		cgroupv2ID: cgroupv2ID,
 	}, nil
 }
 
-// hashString is a helper function for LRUs that use string as a key.
-// Xxh3 turned out to be the fastest hash function for strings in the FreeLRU benchmarks.
-// It was only outperformed by the AES hash function, which is implemented in Plan9 assembly.
-func hashString(s string) uint32 {
-	return uint32(xxh3.HashString(s))
-}
-
 func (r *OTLPReporter) SupportsReportTraceEvent() bool { return true }
-
-// ReportTraceEvent enqueues reported trace events for the OTLP reporter.
-func (r *OTLPReporter) ReportTraceEvent(trace *libpf.Trace, meta *TraceEventMeta) {
-	traceEventsMap := r.traceEvents.WLock()
-	defer r.traceEvents.WUnlock(&traceEventsMap)
-
-	var extraMeta any
-	if r.config.ExtraSampleAttrProd != nil {
-		extraMeta = r.config.ExtraSampleAttrProd.CollectExtraSampleMeta(trace, meta)
-	}
-
-	containerID, err := libpf.LookupCgroupv2(r.cgroupv2ID, meta.PID)
-	if err != nil {
-		log.Debugf("Failed to get a cgroupv2 ID as container ID for PID %d: %v",
-			meta.PID, err)
-	}
-
-	key := samples.TraceAndMetaKey{
-		Hash:           trace.Hash,
-		Comm:           meta.Comm,
-		Executable:     meta.Executable,
-		ApmServiceName: meta.APMServiceName,
-		ContainerID:    containerID,
-		Pid:            int64(meta.PID),
-		ExtraMeta:      extraMeta,
-	}
-
-	if events, exists := (*traceEventsMap)[key]; exists {
-		events.Timestamps = append(events.Timestamps, uint64(meta.Timestamp))
-		(*traceEventsMap)[key] = events
-		return
-	}
-
-	(*traceEventsMap)[key] = &samples.TraceEvents{
-		Files:              trace.Files,
-		Linenos:            trace.Linenos,
-		FrameTypes:         trace.FrameTypes,
-		MappingStarts:      trace.MappingStart,
-		MappingEnds:        trace.MappingEnd,
-		MappingFileOffsets: trace.MappingFileOffsets,
-		Timestamps:         []uint64{uint64(meta.Timestamp)},
-	}
-}
 
 // ReportFramesForTrace is a NOP for OTLPReporter.
 func (r *OTLPReporter) ReportFramesForTrace(_ *libpf.Trace) {}
@@ -216,49 +160,6 @@ func (r *OTLPReporter) FrameKnown(frameID libpf.FrameID) bool {
 		_, known = (*frameMap)[frameID.AddressOrLine()]
 	}
 	return known
-}
-
-// FrameMetadata accepts metadata associated with a frame and caches this information.
-func (r *OTLPReporter) FrameMetadata(args *FrameMetadataArgs) {
-	fileID := args.FrameID.FileID()
-	addressOrLine := args.FrameID.AddressOrLine()
-
-	log.Debugf("FrameMetadata [%x] %v+%v at %v:%v",
-		fileID, args.FunctionName, args.FunctionOffset,
-		args.SourceFile, args.SourceLine)
-
-	if frameMapLock, exists := r.pdata.Frames.GetAndRefresh(fileID,
-		pdata.FramesCacheLifetime); exists {
-		frameMap := frameMapLock.WLock()
-		defer frameMapLock.WUnlock(&frameMap)
-
-		sourceFile := args.SourceFile
-		if sourceFile == "" {
-			// The new SourceFile may be empty, and we don't want to overwrite
-			// an existing filePath with it.
-			if s, exists := (*frameMap)[addressOrLine]; exists {
-				sourceFile = s.FilePath
-			}
-		}
-
-		(*frameMap)[addressOrLine] = samples.SourceInfo{
-			LineNumber:     args.SourceLine,
-			FilePath:       sourceFile,
-			FunctionOffset: args.FunctionOffset,
-			FunctionName:   args.FunctionName,
-		}
-		return
-	}
-
-	v := make(map[libpf.AddressOrLineno]samples.SourceInfo)
-	v[addressOrLine] = samples.SourceInfo{
-		LineNumber:     args.SourceLine,
-		FilePath:       args.SourceFile,
-		FunctionOffset: args.FunctionOffset,
-		FunctionName:   args.FunctionName,
-	}
-	mu := xsync.NewRWMutex(v)
-	r.pdata.Frames.Add(fileID, &mu)
 }
 
 // ReportHostMetadata enqueues host metadata.
